Move derive_generic stuff out
